rename umap.py to dimensionality_reduction.py
