+++ uptrain/core/classes/visuals/__init__.py
@@ -1,3 +1,3 @@
 from .abstract_visual import AbstractVisual
-from .umap import DimensionalityReduction, UMAP_PRESENT
+from .dimensionality_reduction import DimensionalityReduction, UMAP_PRESENT
 from .shap import Shap, SHAP_PRESENT
